[GSF] Syncing Fork

+++ build_tools/build_iOS.sh
@@ -15,6 +15,11 @@ CWD=$PWD
 function build_fat_lib(){
 
     local lib_names=$(cd ./install/$1/iOS/${IOS_ARCHS%% *}/lib; ls *.a)
+    if  [ ! -n "$lib_names" ] ;then
+        echo "break create $1 fat"
+        return
+    fi
+
     local fat_dir=install/$1/iOS/fat/lib
     rm -rf ${fat_dir}
     mkdir -p  ${fat_dir}
@@ -43,7 +48,7 @@ function create_cmake_config(){
 #   echo "find_library(COREVIDEO CoreVideo)" >> $CONFIG_FILE
 #   echo "find_library(COREFOUNDATION CoreFoundation)" >> $CONFIG_FILE
    echo -n "set(SRC_LIBRARIES ${SRC_LIBRARIES}" >> $CONFIG_FILE
-   if [[ "${CURL_SSL_USE_NATIVE}" == "TRUE" ]];then
+   if [[ "${SSL_USE_NATIVE}" == "TRUE" ]];then
        echo -n ' ${SECURITY}' >> $CONFIG_FILE
    fi
    echo -n ' ${AUDIO_TOOL_BOX}' >> $CONFIG_FILE
@@ -66,7 +71,7 @@ function build_shared_framework(){
     export CPU_FLAGS=
     export LDFLAGS=
 
-    local support_libs="fdk-aac x264 curl openssl librtmp cares"
+    local support_libs="fdk-aac x264 curl openssl librtmp cares dav1d"
 
     SRC_LIBRARIES_DIR="$CWD/install/ffmpeg/iOS/fat/lib"
 
@@ -79,6 +84,10 @@ function build_shared_framework(){
         fi
     done
 
+    if [ -d "${DAV1D_EXTERNAL_DIR}/iOS/fat" ];then
+        SRC_LIBRARIES_DIR="$SRC_LIBRARIES_DIR ${DAV1D_EXTERNAL_DIR}/iOS/fat/lib"
+        SRC_LIBRARIES="$SRC_LIBRARIES dav1d"
+    fi
 
  #   $CWD/install/openssl/iOS/fat/lib"
     cd ./install/ffmpeg/iOS/
